[INTERPRETER] * NEW: A new interpreter API for setting the endianness of a stream. * NEW: Another new interpreter API for making a stream blocking or not.
[GB.FORM.MDI]
* NEW: Remove the useless toolbar handle.

[GB.NET]
* BUG: Fix the UdpSocket implementation.
* NEW: UdpSocket now uses an internal buffer, so that successive READs on 
  it work as expected. Beware that the remaining data is lost the next time
  the UdpSocket raises its Read event.



git-svn-id: svn://localhost/gambas/trunk@1669 867c0c6c-44f3-4631-809d-bfa615b0a4ec

+++ gb.net/src/CSocket.h
@@ -45,8 +45,8 @@ extern GB_STREAM_DESC SocketStream;
 
 #endif
 
-void CSocket_CallBack(int t_sock,int type,long lParam);
-void CSocket_CallBackConnecting(int t_sock,int type,long lParam);
+void CSocket_CallBack(int t_sock,int type,intptr_t lParam);
+void CSocket_CallBackConnecting(int t_sock,int type,intptr_t lParam);
 
 typedef  struct
 {
